Upgrade to the latest luci/gae version.
This adjusts to many incompatible changes in the underlying library. The good
news is that the newer luci/gae fixes several shortcomings of the testing
infrastructure, deflaking the test suite and making it more realistic.

This change requires the pending fix to https://github.com/luci/gae/issues/66 to
work.